Add files via upload
You're present at point 'A' which is the top left cell of an M*N matrix. You destination is point 'B', which is the bottom right cell of the same matrix. Your task is to find the total number of unique paths from point 'A' to point 'B'. In other words, you will be given the dimensions of the matrix as integers 'M' and 'N'. You task is to find the total number of unique paths from the cell MATRIX[0][0] to MATRIX[M-1][N-1].

To traverse in the matrix, you can either move Right or Down at each step. For example, in a given point at MATRIX[i][j], you can move to either MATRIX[i+1][j] or MATRIX[i][j+1].