Initial commit
Given a 'N' * 'M' maze with obstacles, count and return the number of unique paths to reach the right-bottom cell from the top-left cell. A cell in the given maze has a value '-1' if it a blockage or dead-end, else 0. From a given cell, we are allowed to move to cells (i+1. j) and (i, j+1) only. Since the answer can be large, print it in modulo 10^9 + 7.